separated endpoints to files
--having everything live in main is bad, actually
