+++ app/core/deps.py
@@ -8,7 +8,7 @@ from pydantic import BaseModel
 from app.core.security import SECRET_KEY, ALGORITHM, ACCESS_TOKEN_EXPIRE_MINUTES,authenticate_user, create_access_token
 from app.model.user import User
 from app.model.token import TokenData, Token
-from app.main import get_user
+from app.api.user import get_user
 
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
